+++ iotdb-core/datanode/src/main/java/org/apache/iotdb/db/pipe/connector/payload/legacy/DeletionPipeData.java
@@ -21,8 +21,8 @@
 package org.apache.iotdb.db.pipe.connector.payload.legacy;
 
 import org.apache.iotdb.commons.exception.IllegalPathException;
-import org.apache.iotdb.db.pipe.connector.protocol.legacy.loader.DeletionLoader;
-import org.apache.iotdb.db.pipe.connector.protocol.legacy.loader.ILoader;
+import org.apache.iotdb.db.pipe.receiver.legacy.loader.DeletionLoader;
+import org.apache.iotdb.db.pipe.receiver.legacy.loader.ILoader;
 import org.apache.iotdb.db.storageengine.dataregion.modification.Deletion;
 import org.apache.iotdb.tsfile.utils.ReadWriteIOUtils;
